Improved comments, included Cov. Matrix output, PrepareEnd, EXIT_AT.
We introduced a lot of new comments in the code and organized it.
Also, we included the possibility of outputting the data with a versatile
function that outputs Eigenvalues, Eigenvectors and Cov. Matrix.
So far the output was only tested (and not checked) with Cov. Matrix.
We also made the code more "flask-like" by introducing the EXIT_AT
possibility and the PrepareEnd function.