Computing radial part of Noise covariance matrix.
This version computes the radial part of the noise covariance matrix
using FFTW. We are currently testing it and it seems ok.

We also made the code transform the original inputs such as completeness
maps and redshift distribution of galaxies into the functions we use
in the noise covariance matrix. This conversion requires cosmological
formulas such as comoving distance.