Fixed approximation bug when computing fftwIn
exp(log(r)) was returning a value slightly smaller than r and
this was causing a discrepancy between the ZZr2 with rWin branch and
this branch. By adding a error tolerance when comparing r's to decide
wether or not to set the noise term to zero outside the radial window,
we removed this discrepancy.