ZZr2 implemented and contributing to Cov Matrix
We implemented the ZZr2 transform, its results are still with an
factor of 0.955 in W-w when compared to Mathematica result, but
apart from that the results seem correct. We made some cleanups
like calling Q-q by qq and leaving q for q. I also introduced the
ZZR2_OUT output and included missing EXIT_ATs.

TODO:
- check why Mathematica and zzr2 trafo disagree on W-w by 0.955 factor.
  One idea is to check its implications in the formulas written on paper.
- optimize computation of cov. matrix since angular part is always the same
  for different W-w but is being computed every time.
- compare results with simulations.