Ang and rad cov. matrix independent, results same as together
We found out that there was a mistake in the previous commit, that the blocks
corresponding to different radial modes were not including the top
part because we only compute the bottom part of the angular cov. matrix.
In this version we fixed this issue and verified that its results are the
same as the previous version in which we computed everything together.