Began coding radial part of noise
In this version we are in the middle of coding the radial part of the noise
covariance matrix and therefore the full matrix. We added a few new parameters
such as SELEC_SCALE and intermediate outputs MASK_OUT, COVCONST_OUT, ANGCOV_OUT.
We realised that we needed to compute a constant part related to not subtracting the
mean. This was implemented, which involved computing a binary mask from the noise
map.

We compared the results from this code (the theoretical prediction of the Zwlm
covariance matrix) to simulations and they were very off. Actually the angular
part seems to be wrong and different from before. We need to investigate this.