Added Romberg method to ZZR2 transform and the ZZR2_METHOD keyword
This version allows the user to choose between two methods of computing the
radial part of the noise covariance matrix [the Z_w(r) Z*_W(r) / n(r) transform
or integral], the original FFT one and the new Romberg integral. For the latter,
we included a new version of the qromb routine that allows tabulated functions
that are later interpolated to be integrated. This version required a slight
decrease in precision for qromb to work. The results were compared to Mathematica's
integration, and they agree.

TODO: note that the FFT and Romberg method still don't agree among themselves
just like the Mathematica integration and FFT case (and none agrees perfectly
with the results from simulations). I should understand this.